chore: prefactor move dashboard alerts into component
